rest/requirements.txt
alembic==1.8.0
bcrypt==3.2.2
fastapi-utils==0.2.1
Jinja2==2.11.3
psycopg2-binary==2.9.3
pydantic[email]
MarkupSafe==2.0.1
pydantic==1.9.1
PyGithub==1.55
PyJWT==2.4.0
pytzdata==2020.1
PyYAML==5.4.1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.9
SQLAlchemy-Utils==0.38.2
uvicorn==0.18.1
uvicorn[standard]==0.18.1
python-dotenv==0.20.0
passlib==1.7.4
tomli==2.0.1
tomli_w==1.0.0
cryptography==39.0.1
aiohttp==3.8.3